requirements.txt
torch>=2.0.0
transformers>=4.36.0
accelerate>=0.25.0
pillow>=10.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.4.2
requests>=2.31.0
python-multipart>=0.0.6
bitsandbytes>=0.41.0
# flash-attn is optional and can cause build issues
# flash-attn>=2.3.0
datasets>=2.14.0
peft>=0.6.0
trl>=0.7.4
tqdm>=4.65.0
pandas>=2.0.0
